requirements.txt
pathway==0.6.0
pandas==2.1.0
numpy==1.24.0
sentence-transformers==2.2.2
torch==2.0.0
transformers==4.35.0
scikit-learn==1.3.0
pyyaml==6.0.1
tqdm==4.66.0